VirtualTable.map() uses MapperFactory instead of Map function
The MapperFactory creates Runnable mappers that always use the same
inputs/outputs. The output ColumnSchema was also moved to MapperFactory.

MapperFactory contains static convenience methods to construct MapperFactories
for standard java.util.function interfaces, for example DoubleBinaryOperator,
etc. (This is proof-of-concept, support for more functional interfaces should
be added if this is useful).