AP-18625: implement comp graph optimization moveSlicesBeforeAppends()
This targets VirtualTables that append several source tables and then slice
the result. The new optimization pushes the "slice" from after the "append"
to before the append (on all incoming branched). This then allows the
mergeSliceToSource() optimization to push the "slice" further into the
source. So for example, for Arrow tables, the slice is then directly done
on the Arrow columns.