AP-21532: RagGraph copying
This copy() method makes a copy of the SPEC graph (before access tracing and optimization).
It will currently not work to make a full copy of a RagGraph after access tracing and optimization, or intermediate stages.

It may be possible later to extend this to copy partially optimized graphs, but that will require additional pruning of AccessIds that
have no remaining consumers, and pruning of Access-producing nodes that have no produced accesses left.