AP-22734: Use ReturnResult for column+var type resolution
This allows the caller to provide a helpful error message if the column access
or variable access is not applicable. It is not applicable if the column
or variable is missing, or if they have a type that is not supported by
expressions. With Optional, we could not differentiate between those
reasons.

AP-22734 (Expressions offer local date columns to use, but access them as long value)