AP-22734: Remove DATA_SPEC_TO_EXPRESSION_TYPE mapper
It resolves expression types based on the internal representation of data in a table.
However, some types are represented as data that is supported in expressions but the
type should not be (e.g. LocalDate is represented as Long).

AP-22734 (Expressions offer local date columns to use, but access them as long value)