AP-23851: always make a copy of graph in appendSelection()
... even if selection == Selecton.allSelected()

This fixes a bug in CapRowAccessible.createCursorData:
It can happen that createCursorData is called with two different
selections from two threads. Each of the threads then calls
appendSelection() and optimizes the resulting graph.
Previously, if one of the selections was allSelected(), then the
respective thread would optimize the original graph, while the other
thread would try to make a copy at the same time.